fix: Reestruturação de arquivos

+++ src/Travels.Application/Interfaces/Repositories/IRouteRepository.cs
@@ -1,4 +1,6 @@
 using Travels.Domain.Entities;
+using Travels.Application.Common;
+using Travels.Application.UseCases.GetCheapestRoute;
 
 namespace Travels.Application.Interfaces.Repositories;
 
@@ -6,7 +8,7 @@ public interface IRouteRepository
 {
     Task AddAsync(Route route);
     Task<List<Route>> GetAllAsync();
-    Task<Route?> GetByIdAsync(int id);
-    Task DeleteAsync(int id);
+    Task<Route?> GetByIdAsync(long id);
+    Task DeleteAsync(long id);
     Task UpdateAsync(Route route);
 }
